requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-phonenumber-field==4.0.0
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
isort==4.3.21
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
phonenumbers==8.12.4
Pillow==7.1.2
psycopg2==2.8.5
pygame==1.9.6
Pygments==2.6.1
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.8
virtualenv==20.0.20
whitenoise==5.1.0
wrapt==1.12.1
